backend/requirements.txt
# AI University Backend Dependencies
# Все библиотеки стандартные, без vendor lock-in

# === Web Framework ===
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# === Database (vendor-agnostic PostgreSQL) ===
sqlmodel==0.0.22
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.2

# === OpenAI (стандартный SDK с AsyncOpenAI) ===
openai==1.51.0

# === S3 Storage (asynchronous, S3-compatible) ===
aioboto3==13.1.1

# === Telegram Bot (aiogram 3.x) ===
aiogram==3.13.1
redis==5.0.8

# === Utilities ===
phonenumbers==8.13.29

# === Authentication (JWT) ===
PyJWT==2.9.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# === Configuration ===
pydantic-settings==2.5.2
python-dotenv==1.0.1

# === Logging ===
loguru==0.7.2

# === Utilities ===
httpx==0.27.2
python-dateutil==2.9.0
pytz==2024.2
APScheduler==3.10.4

# === Type Validation ===
email-validator>=2.0.0
tiktoken==0.7.0
